refactor(P2.3): move clause extractors into render_plan/clause_extractors.rs
Phase-2 §5.1 module move. Extracted the pure clause extractors that remained in
plan_builder_utils.rs verbatim to a new render_plan/clause_extractors.rs:
extract_having / extract_order_by / extract_limit / extract_skip +
extract_sorted_properties (old lines 1467-1560). `pub(crate) use` re-exports
left in plan_builder_utils so all `super::plan_builder_utils::extract_*` call
sites and properties_builder's direct import keep resolving during the
transition. Zero logic edits.

Stale-premise correction: §5.1 named an 8-function group
(filters/from/group_by/having/order_by/limit/skip/distinct) but
extract_filters/from/group_by/distinct had already migrated to
filter_builder.rs/group_by_builder.rs via incremental work, so P2.3's real
scope was the 5-function remainder. §5.1 table + §9 checklist + PRIORITIES P-3
updated to match.

Gates: 1612 lib + 529 integration (incl. corpus_sweep + sql_golden) + 1 ratchet
+ 7 unit, 0 failures; byte-identical (no golden churn, corpus snapshot
unchanged, ratchet net-zero); fmt/clippy clean. Moved-fn bodies verified
byte-identical vs main.
